# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alivio Search Cloud customer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface CurrentUser {
  id: string;
  clerkId: string;
  email: string;
  fullName?: string | null;
  avatarUrl?: string | null;
  createdAt: string;
}

export interface Business {
  id: string;
  ownerId: string;
  name: string;
  slug: string;
  industry?: string | null;
  websiteUrl?: string | null;
  description?: string | null;
  brandColor?: string | null;
  logoUrl?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateBusinessBody {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name?: string;
  /** @maxLength 200 */
  industry?: string;
  /** @maxLength 500 */
  websiteUrl?: string;
  /** @maxLength 2000 */
  description?: string;
  /** @maxLength 32 */
  brandColor?: string;
  /** @maxLength 500 */
  logoUrl?: string;
}

export type AgentSummaryMode =
  (typeof AgentSummaryMode)[keyof typeof AgentSummaryMode];

export const AgentSummaryMode = {
  structured: "structured",
  text: "text",
} as const;

export interface AgentSummary {
  id: string;
  name: string;
  description: string;
  mode: AgentSummaryMode;
}

export type RunAgentRequestInput = { [key: string]: unknown };

export interface RunAgentRequest {
  input: RunAgentRequestInput;
}

export type AgentRunInput = { [key: string]: unknown } | null;

export type AgentRunOutput = { [key: string]: unknown } | null;

export interface AgentRun {
  id: string;
  businessId: string;
  userId: string;
  agentSlug: string;
  status: string;
  input?: AgentRunInput;
  output?: AgentRunOutput;
  errorMessage?: string | null;
  tokensUsed?: number | null;
  startedAt: string;
  completedAt?: string | null;
}

export type ServiceStatusValue =
  (typeof ServiceStatusValue)[keyof typeof ServiceStatusValue];

export const ServiceStatusValue = {
  ok: "ok",
  configured: "configured",
  not_configured: "not_configured",
  error: "error",
} as const;

export type ServiceStatusMessaging = {
  twilio: ServiceStatusValue;
  whatsapp: ServiceStatusValue;
};

export interface ServiceStatus {
  api: ServiceStatusValue;
  database: ServiceStatusValue;
  ai_provider: ServiceStatusValue;
  stripe: ServiceStatusValue;
  cal_link: ServiceStatusValue;
  messaging: ServiceStatusMessaging;
}

export type SmokeTestResultStatus =
  (typeof SmokeTestResultStatus)[keyof typeof SmokeTestResultStatus];

export const SmokeTestResultStatus = {
  pass: "pass",
  fail: "fail",
  not_configured: "not_configured",
} as const;

export interface SmokeTestResult {
  id: string;
  name: string;
  status: SmokeTestResultStatus;
  latencyMs: number;
  error?: string | null;
  preview?: string | null;
}

export interface SmokeTestReport {
  runAt: string;
  passed: number;
  failed: number;
  notConfigured: number;
  results: SmokeTestResult[];
}

export type ListAgentRunsParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
  agentId?: string;
};