# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Alivio Search Cloud customer API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication and current user
  - name: businesses
    description: Business workspace operations
  - name: agents
    description: AI agent execution
  - name: agent-runs
    description: Agent run history
  - name: status
    description: Service status
  - name: admin
    description: Admin-only endpoints
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /status:
    get:
      operationId: getStatus
      tags: [status]
      summary: Service status overview
      description: Returns the configured/connected status of every backing service. Never returns secret values.
      responses:
        "200":
          description: Status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceStatus"
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Current authenticated user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /businesses/current:
    get:
      operationId: getCurrentBusiness
      tags: [businesses]
      summary: Current business workspace
      responses:
        "200":
          description: Current business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCurrentBusiness
      tags: [businesses]
      summary: Update the current business workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBusinessBody"
      responses:
        "200":
          description: Updated business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Business"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all available agents
      responses:
        "200":
          description: Agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agents/{id}/run:
    post:
      operationId: runAgent
      tags: [agents]
      summary: Run an agent
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunAgentRequest"
      responses:
        "200":
          description: Agent run record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agent-runs:
    get:
      operationId: listAgentRuns
      tags: [agent-runs]
      summary: List recent agent runs for the current business
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: agentId
          schema:
            type: string
      responses:
        "200":
          description: Agent runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRun"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /agent-runs/{id}:
    get:
      operationId: getAgentRun
      tags: [agent-runs]
      summary: Single agent run
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/agents/smoke-test:
    post:
      operationId: smokeTestAgents
      tags: [admin]
      summary: Smoke-test every agent (admin only)
      responses:
        "200":
          description: Smoke test results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmokeTestReport"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — not an admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    CurrentUser:
      type: object
      properties:
        id:
          type: string
        clerkId:
          type: string
        email:
          type: string
        fullName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, email, createdAt]
    Business:
      type: object
      properties:
        id:
          type: string
        ownerId:
          type: string
        name:
          type: string
        slug:
          type: string
        industry:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        brandColor:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, ownerId, name, slug, createdAt, updatedAt]
    UpdateBusinessBody:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        industry:
          type: string
          maxLength: 200
        websiteUrl:
          type: string
          maxLength: 500
        description:
          type: string
          maxLength: 2000
        brandColor:
          type: string
          maxLength: 32
        logoUrl:
          type: string
          maxLength: 500
    AgentSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        mode:
          type: string
          enum: [structured, text]
      required: [id, name, description, mode]
    RunAgentRequest:
      type: object
      properties:
        input:
          type: object
          additionalProperties: true
      required: [input]
    AgentRun:
      type: object
      properties:
        id:
          type: string
        businessId:
          type: string
        userId:
          type: string
        agentSlug:
          type: string
        status:
          type: string
        input:
          type: object
          additionalProperties: true
          nullable: true
        output:
          type: object
          additionalProperties: true
          nullable: true
        errorMessage:
          type: string
          nullable: true
        tokensUsed:
          type: integer
          nullable: true
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, businessId, userId, agentSlug, status, startedAt]
    ServiceStatusValue:
      type: string
      enum: [ok, configured, not_configured, error]
    ServiceStatus:
      type: object
      properties:
        api:
          $ref: "#/components/schemas/ServiceStatusValue"
        database:
          $ref: "#/components/schemas/ServiceStatusValue"
        ai_provider:
          $ref: "#/components/schemas/ServiceStatusValue"
        stripe:
          $ref: "#/components/schemas/ServiceStatusValue"
        cal_link:
          $ref: "#/components/schemas/ServiceStatusValue"
        messaging:
          type: object
          properties:
            twilio:
              $ref: "#/components/schemas/ServiceStatusValue"
            whatsapp:
              $ref: "#/components/schemas/ServiceStatusValue"
          required: [twilio, whatsapp]
      required:
        [api, database, ai_provider, stripe, cal_link, messaging]
    SmokeTestResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [pass, fail, not_configured]
        latencyMs:
          type: integer
        error:
          type: string
          nullable: true
        preview:
          type: string
          nullable: true
      required: [id, name, status, latencyMs]
    SmokeTestReport:
      type: object
      properties:
        runAt:
          type: string
          format: date-time
        passed:
          type: integer
        failed:
          type: integer
        notConfigured:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/SmokeTestResult"
      required: [runAt, passed, failed, notConfigured, results]