# lib/api-zod/src/generated/types/agentRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alivio Search Cloud customer API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentRunInput } from "./agentRunInput";
import type { AgentRunOutput } from "./agentRunOutput";

export interface AgentRun {
  id: string;
  businessId: string;
  userId: string;
  agentSlug: string;
  status: string;
  input?: AgentRunInput;
  output?: AgentRunOutput;
  errorMessage?: string | null;
  tokensUsed?: number | null;
  startedAt: Date;
  completedAt?: Date | null;
}