# lib/api-zod/src/generated/types/agentSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alivio Search Cloud customer API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentSummaryMode } from "./agentSummaryMode";

export interface AgentSummary {
  id: string;
  name: string;
  description: string;
  mode: AgentSummaryMode;
}