# lib/api-zod/src/generated/types/runAgentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alivio Search Cloud customer API
 * OpenAPI spec version: 0.1.0
 */
import type { RunAgentRequestInput } from "./runAgentRequestInput";

export interface RunAgentRequest {
  input: RunAgentRequestInput;
}