# lib/api-zod/src/generated/types/serviceStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alivio Search Cloud customer API
 * OpenAPI spec version: 0.1.0
 */
import type { ServiceStatusMessaging } from "./serviceStatusMessaging";
import type { ServiceStatusValue } from "./serviceStatusValue";

export interface ServiceStatus {
  api: ServiceStatusValue;
  database: ServiceStatusValue;
  ai_provider: ServiceStatusValue;
  stripe: ServiceStatusValue;
  cal_link: ServiceStatusValue;
  messaging: ServiceStatusMessaging;
}