# lib/api-zod/src/generated/types/serviceStatusValue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alivio Search Cloud customer API
 * OpenAPI spec version: 0.1.0
 */

export type ServiceStatusValue =
  (typeof ServiceStatusValue)[keyof typeof ServiceStatusValue];

export const ServiceStatusValue = {
  ok: "ok",
  configured: "configured",
  not_configured: "not_configured",
  error: "error",
} as const;