# lib/api-zod/src/generated/types/smokeTestReport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alivio Search Cloud customer API
 * OpenAPI spec version: 0.1.0
 */
import type { SmokeTestResult } from "./smokeTestResult";

export interface SmokeTestReport {
  runAt: Date;
  passed: number;
  failed: number;
  notConfigured: number;
  results: SmokeTestResult[];
}