# lib/api-zod/src/generated/types/smokeTestResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Alivio Search Cloud customer API
 * OpenAPI spec version: 0.1.0
 */
import type { SmokeTestResultStatus } from "./smokeTestResultStatus";

export interface SmokeTestResult {
  id: string;
  name: string;
  status: SmokeTestResultStatus;
  latencyMs: number;
  error?: string | null;
  preview?: string | null;
}